Add types for documentation
Also tidied up the documentation in certain places. The types do not
provide much safety. This is mostly because of the generality of pattern
matching working with arbitrary data. Some areas could be theoretically
improved, but Lua LS is understandably not typescript. Particularly,
it was difficult to overload `match_cond` and to provide documentation for
the matchers.