Change __eq's interaction with unification
Unification of tables now proceed by first trying to unify the values
and variables corresponding to a case's keys. If successful, the
inner variables are unified as before. If this simple unification fails,
an equality test is used with the object on the left-hand side. This
invokes the __eq metamethod of the object, if it exists.
If the equality test succeeds, then the original substitution hash is
returned, and used in the rest of the computation. If it fails, then
unificaiton fails as before.

Some test cases are added to exemplify this behaviour.